Added rendering contexts, shaders and comments.

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -21,7 +21,7 @@ namespace Hazel
 	ImGuiLayer::~ImGuiLayer()
 	{
 	}
-
+	
 	void ImGuiLayer::OnAttach()
 	{
 		// Setup Dear ImGui context
@@ -73,7 +73,7 @@ namespace Hazel
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
 
 		// Rendering
 		ImGui::Render();
